Restrict VP checkcast tenured alignment logic to loadaddr
If the type child of the checkcast node is a static load (rather than
loadaddr), then the static address will be incorrectly interpreted as a
class pointer.

This change is just a precaution. It's unlikely that it will be possible
for a static load to appear in this position for some time, even for a
checkcast node whose cast type is dynamic.