TLH Alignment for Spanning Sweep Chunks
Normally, a TLH can span multiple sweep chunks without any issues.
However, a batch marked TLH (for SATB) is required to be contained
within a single sweep chunk. Otherwise, this is problematic for sweep as
sweep must determine the last obj in a chunk. With a batch marked TLH
spanning multiple chunks, the TLH in the fist chunk will only be
partially contained. As a result, all mark bits will be set to the end
of the chunk resulting in the last live obj in chunk being
undiscoverable. Hence, we must align the TLHs to sweep chunk boundaries
when a pool freeEntry spans multiple chunks during TLH allocation.

To do so, we first determine the chunk boundaries containing the
freeEntry baseAddr. If freeEntry topAddr is past the chunk top boundary
(the freeEntry spans multiple chunks) then we clip part of the freeEntry
which overflows into next chunk.

Signed-off-by: Salman Rana <salman.rana@ibm.com>