Improve container detection
Add logic to detect the presence of container files for both cgroup v1
and v2. If these files are not present, resort to the original method
(which will not work if namespaces are enabled in the container).
Also, the cgroup mount point is renamed as both cgroup v1 and v2 are
conventionally mounted at the same point.

Issue: #1281
Signed-off-by: Eric Yang <eric.yang@ibm.com>