Consider parm invariance when merging def nodes for GVP
It is possible in DLT compiles for a store to a parm slot to be missed
by GVP analysis because it appears in a part of the method that occurred
prior to the DLT entry point and hence IL trees were never generated
for it.  It may then be possible for GVP to conclude that the only def
to the parm slot is at method entry and then assume that the characteristics
of the parm slot at entry (such as its type) are how uses of the parm
slot should be constrained.  This is incorrect in DLT compiles and the
analysis needs to be made more conservative and assume parms are variant.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>