Choose which child to evaluate correctly
Using INT_MAX as the initial best evaluation priority results in each
node receiving a priority of 0 in whichChildToEvaluate, because no
childPriority will ever be greater than the INT_MAX value. Using
INT_MIN instead ensures that bestPriority corresponds to the deepest
child depth in the tree, and updates bestChild and nodePriority
accordingly.