AArch64: Do not skip l2a node if the MemoryReference has UnresolvedDataSnippet
`populateMemoryReference` function skips `l2a` node if the node's reference
count is 1 and node does not have a register.
Skipping it makes the base register of the memory reference
a non-collected reference register. This is incorrect but harmless in normal
cases. However, if the memory reference has a UnresolvedDataSnippet,
the first generated instruction would be a branch to the snippet. The resolution
helper is called fron the snippet, which may trigger GC. Because GC does not
know about the base register, the value of the base register would not be
updated correctly.
This commit changes `populateMemoryReference` to evaluate `l2a` node if
the memory reference has a UnresolvedDataSnippet.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>