Restrict checkcast and profiled guard coalescer cold merge special case
...to instances where the conditional on the passing side of the guard
does not evaluate any indirect loads. If the conditional does evaluate
an indirect load, the transformation must be abandoned just as though an
indirect load were found in an earlier tree.

The failure mode against which this commit protects is theoretical and
has not been observed. It's unclear whether it's even possible to cause
an indirect load to appear in such a position. Nonetheless, it's prudent
to check.