Add genLoadProfiledClassAddressConstant helper in Z codegen
genLoadAddressConstant only adds instructions from aconst and aloadi
nodes to the staticPICSites lists. Profiled class address constants are
not found in aconst or aloadi nodes, but need to be treated as aconst
nodes. This commit adds the genLoadProfiledClassAddressConstant helper,
to wrap the class address constant value in a dummy aconst node before
passing it to genLoadAddressConstant.

Signed-off-by: Spencer Comin <spencer.comin@ibm.com>