Allow JitBuilder Struct types to define fields not at the end
Currently, fields are added at the end of a struct and cannot
be defined earlier in a struct. To make it easier to define
unions in JitBuilder2, this change allows fields to be defined
that reside not at the end of the struct. While unions have
never been strictly defined by JitBuilder, they have never
really supported union punning (this commit dooesn't change
that, but it does make it easier for someone to attempt to
rely on punning when it isn't supported). This change loosens
field definition rules, so it will not negatively affect any
existing users of struct types.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>