Give destination string length to strncpy
Recent clang compilers will warn if the source string length
is provided to strncpy, so JitBuilder cannot be built due
to this warning. Also updated the use of the brittle "64"
to use a named constant and increased the size to 128 since
there have been complaints in the past about 64 not being
large enough.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>