Get opendir to use the cwd if the corepath is blank
On AIX, if the corepath provided to appendCoreName() is blank, then opendir() will fail. However the core may have been written to the current working directory and it will be missed. If corepath is blank, provide "." to opendir().

Also add some extra messages that will come out when DUMP_DEBUG is defined, and fix a typo.

Signed-off-by: David McCann mccannd@uk.ibm.com