Provide a way to mark cold the taken side of a profiled guard
There can be circumstances under which it is technically necessary to
generate a run-time test, but where the failure of that run-time test is
known in advance to be an exceedingly rare corner case. In such a
circumstance, the taken side should be marked cold, which can now be
arranged by setting _forceTakenSideCold to true.