Stop sharing TR_VirtualGuard between multiple guards with the same BCI
Previously, the TR_VirtualGuard corresponding to a guard node was
identified by a combination of its BCI and its guard kind. The guard
kind was partially erased so that some of the guard kinds (e.g. HCR and
OSR guards) were considered significant, while most regular virtual
guard types were not distinguished from one another. So when given an
HCR guard node, findVirtualGuardInfo() would look specifically for a
TR_VirtualGuard with kind TR_HCRGuard, but when given, say, a profiled
guard, it would turn up the first guard it found with matching BCI and
without a 'significant' guard kind, which could be e.g. a nonoverridden
guard instead.

This old way of correlating guard nodes each with their corresponding
TR_VirtualGuard was troublesome because of two possible kinds of
transformations, both of which must happen to cause an issue:

1. Duplication of the guard node. This can happen deliberately (as in
   versioner), or it can happen simply as part of duplication of the
   containing block.

2. (a) Removal of guards, or
   (b) replacement of guards by others, or
   (c) modification of TR_VirtualGuard, e.g. removing merged guards.

In 2(a), it makes sense to remove the TR_VirtualGuard along with the
guard node so that we aren't left with orphan TR_VirtualGuard instances,
and some code paths have been doing so. However, if there are other
copies of the guard node, those share the TR_VirtualGuard, and removing
it leaves them without one. Many callers of findVirtualGuardInfo() check
defensively to make sure that the TR_VirtualGuard was found, but it's
not necessarily always obvious what to do when the TR_VirtualGuard is
missing. Ideally, a virtual guard node would always have one. This case
could be solved by keeping the TR_VirtualGuard even when all of its
corresponding nodes have been removed, or by reference counting it, but
that won't work for 2(b) or 2(c).

In 2(b), we replace e.g. a profiled guard with a hierarchy guard. If
there are other copies of the original profiled guard, then the old
strategy of keying on BCI ensures that there will be a guard node for
which findVirtualGuardInfo() finds a TR_VirtualGuard instance with a
guard kind that differs from the kind expected by the node.

In 2(c), we update the TR_VirtualGuard based on its occurrence at one
particular guard node. If there are other copies of the guard node, then
the update will apply to those as well, even if it shouldn't. I don't
believe it's currently possible for this (2(c)) issue to occur, but it
would be easy to accidentally introduce via a seemingly innocuous
change, e.g. running block splitter before OSR guard insertion.

This commit changes to a strategy where each guard node has its very own
instance of TR_VirtualGuard, which can in a way now be considered to be
part of the guard node itself, like the node extension. Whenever a guard
node is removed or replaced, its TR_VirtualGuard can always be removed
as well, since we know that it isn't shared with any other node. When a
guard node is duplicated, its TR_VirtualGuard must now be duplicated
along with it.

It's no longer necessary to search for the TR_VirtualGuard belonging to
a given guard node, since the guard node now has a pointer to it.

The node flags no longer redundantly specify the guard kind, so there is
no way for the node and the TR_VirtualGuard to disagree on it.

Finally, guard nodes in trace logs now show the test type and indicate
the presence of any merged HCR or OSR guard or inner assumption (though
the particular inner assumptions are still not shown).