Add support for CUDA version 11.7
Remove useless entries for versions 11.n where n is not zero.

For example, version 11.7 for Linux includes these runtime libraries:
  libcudart.so         -> libcudart.so.11.0
  libcudart.so.11.0    -> libcudart.so.11.7.60
  libcudart.so.11.7.60
On Windows, version 11.7 includes this shared library:
  cudart64_110.dll

There is nothing named "libcudart.so.11.7" or "cudart64_117.dll", so
entries for them should not expect find anything. The story is similar
for versions 11.1 through 11.6, inclusive.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>