Version profiled guards with cold targets regardless of profiling
The target block is only marked cold when the guard branch is strongly
expected to fall through. This might be the case for a profiled guard
even when the top profiled receiver type is not overwhelmingly likely,
e.g. if the guard uses a method test and all profiled receivers inherit
the same method.

It's possible to look specifically for such situations, but for the
moment, just don't second-guess the cold marking on the target block.