Don't emit rex prefix on 32-bit
The rex prefix is illegal in 32-bit mode. However,
some VEX and EVEX encoded instructions use REX.W the bit
as an opcode extension bit. In 32-bit mode, if a SIMD
instruction is encoded as a legacy SSE instruction,
and the W bit is set, a rex prefix may be emitted. This
commit prevents that from happening.

Signed-off-by: BradleyWood <bradley.wood@ibm.com>