AArch64: Use vector eor instructions for register exchange
In register assignment, a temporary register is used for
exchanging floating point or vector registers.
However, if a register dependency has all 32 FPRs, the register assigner
cannot find the temporary register and fails abruptly.
Since vector eor instructions can be used for exchanging floating point
or vector registers without a temporary register, this commit changes
the register assigner to generate those instructions for exchanging registers
and removes the use of a temporary register.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>