Rework profiled to nop guard upgrades in VP
For all profiled guards (VFT test or method test), when the guard is not
folded, replace it with a noppable guard if possible.

Replacing a VFT test with a noppable guard is possible when there is a
type bound C for the receiver, and C is the type expected by the VFT
test, and C has not been extended. It's as though C is final (which
would allow folding via a fixed-type constraint), but only until further
notice.

Replacing a method test with a noppable guard is possible when there is
a type bound C for the receiver, and C has the expected method, and the
expected method is not overridden. It's as though the expected method is
final (which would allow folding), but only until further notice.

By default such upgrades are performed only outside of the VM startup
phase, since during startup the VM is still loading classes pretty
frequently, so the non-extended or nonoverridden state is not as
significant. To enable the transformation during startup, set the
environment variable TR_upgradeToNopGuardDuringStartup.

A new query VMEnv::isVMInStartupPhase(TR::Compilation*) is defined to
detect this startup phase. It is intended for downstream projects to
override. The default implementation arbitrarily returns false.

There was pre-existing code that attempted to change profiled guards
into nop guards, but it worked only for VFT tests, and only when the
cold call was a virtual call (although part of the code appeared to be
trying to handle interface calls). The new implementation also works for
method tests and for interface calls, and in the case of method test,
the logic that triggers the transformation is largely shared with other
recent improvements to the handling of method test. As such, this commit
deletes the prior implementation of this concept.