Allow VP to remove/upgrade virtual guards with merged HCR/OSR guards
Allow folding (or for profiled guards, replacement by noppable guards)
to proceed even if the guard in question is merged with an HCR or an OSR
guard. Guards folded to always go to the cold side need no special
handling. Guards folded to always go to the hot side (i.e. to fall
through) have any merged HCR and/or OSR guard split out so that the
virtual guard can be removed. When a profiled guard is replaced by a
noppable guard, the noppable guard has the same merged HCR or OSR guard
as the original profiled guard did, if any.

Currently, this only affects guards that can be folded simply based on
the value numbers or constraints of the children (e.g. mustBeEqual()),
or guards that use VFT test or method test. The logic that folds
nonoverridden guards has canFoldNonOverriddenGuard() check separately
for canBeRemoved(). In principle it should be fine to relax that check
too, but canFoldNonOverriddenGuard() is also used in constrainCall(),
which instead of using removeConditionalBranch() has its own way of
removing the guard. And it needs its own way. For example, it has to do
TR::TransformUtil::removeTree(...), not vp->_curTree->setNode(NULL). It
also might not be able to setUnreachablePath() on the edge. So to extend
this improvement to the elimination of nonoverridden guards, some of
this new logic will need to be factored out into its own function, and
that work is left for the future.