Prevent loading mark bits prematurely during final card cleaning
On ARM M1 processor, we may end up marking an object on one thread/core
that happens to be in an already dirty card, hence skipping to scan it,
while the other thread/core racing to clean the card does not see the
object being marked and skips scanning it. Altogether, we miss to scan
the object.

Normally not possible, but due to some kind of store or load reordering
the latter core sees a stale mark bit.

A fix that appears to work is to prevent the loading of mark bits too
early. It's not completely clear what other loading operation this can
be reorder with... Loading of mark bits is conditional to prior reading
a card, and only proceeding with reading mark bits, if the card is
dirty. Hence, first load operation of card reading is expected to
complete before reading mark bits.

This problem has to be studied more, but for now, this is a
simple fix, relatively non-intrusive, that is tested and works.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>