Fix an assert in OMR::Power::TreeEvaluator::s2iEvaluator()
The s2iEvaluator uses a isLoad() test when deciding between a simple
sign extend or generating a load sequence. The isLoad() test returns
true for both LoadVar and LoadConst cases and therefore we can get
into the load sequence when handling a sconst node child which results
in an assert. This PR will change isLoad() to isLoadVar() so that
the sconst case will use the simple sign extend path.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>