Restrict VP's VFT and method test guard handling to ifacmpne
Guards using ifacmpeq (i.e. guards for which the branch node has been
reversed) are hard to generate and therefore hard to test. Furthermore,
they should arguably be disallowed entirely. Reversed nop guards must
already be disallowed (since nop guards necessarily fall through when
passing and jump when failing), and it's unclear what it would even mean
for a reversed profiled guard to be merged with an HCR or OSR guard.

Requiring ifacmpne here allows some small simplifications to the
existing guard handling logic. It will also simplify the logic for
upcoming improvements to the handling of method test, and it will help
clarify possible future handling of merged guards and profiled to nop
guard upgrades.