Check predecessors of call before considering eliminating guard
In applying constraints for a non-overridden call guard,
constrainIfcmpeqne looks at the guarded call node to see whether the
guard can be eliminated.  That's done using canFoldNonOverriddenGuard
which calls mergeDefConstrants.  That can have the side effect of
creating global constraints.  However, if there are paths to the call
block that do not go through the block that contains the if the
information used to create those global constraints might be incomplete.

This change adds an additional check of whether the block containing the
guarded call has only one predecessor and whether the block containing
the call is actually the target of the branch before checking whether
the guard can be eliminated.  That avoids the possibility of other merge
points between the block containing the if and the block containing the
call, and hence that definitions from paths that do not include the if
might reach the call.