Defend removed nodes during delayed transformations
During globalVP nodes that are added to lists for later processing
may be removed from the trees before doDelayedTransformations is called.
If such nodes are processed the trees can be corrupted and will likely
cause a crash during compilation.

The problem was originally seen with the transformArrayCloneCall, and
has also been demonstrated with transformObjectCloneCall. This PR avoids
the issue for these two functions by testing the reference count of the
call node.

It is likely that other transformations called from doDelayedTransformations
will need similar treatment.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>