Ignore opcodes when attempting to fold array class tests
The previous logic here incorrectly assumed that an aloadi must load
<vft-symbol>, but it could load from elsewhere, e.g. a field of a
reflective object. When loading <classFromJavaLangClass>, the base
object could be constrained with isClassObject() == TR_yes, which would
result in the right type even though it doesn't seem like the previous
logic was intended to take this case into account. But it could also be
constrained with isClassObject() != TR_yes, and with a type bound. In
that case, the type bound describes the type of the object we're loading
from, which is actually as expected by the previous logic, but because
the aloadi loads <classFromJavaLangClass> rather than <vft-symbol>, that
type is not in general related to the result of the load, and VP could
transform the tree incorrectly. More specifically, the test could be
folded to 0 (indicating a non-array type) even though the reflectively
identified type could turn out to be an array type at runtime.

The opcode of the node that produces the class pointer is irrelevant.
Just get the constraint for that node, which always has a consistent
interpretation, instead of trying to guess what getClassType() means
based on multiple possible choices of baseExpression.

This change also generalizes the transformation to cases where the
opcode is neither aloadi nor loadaddr. For example, the class pointer
could come from a direct aload of a temporary, if we have a bound for
the classes that the temporary could possibly point to. As such, there
are no "unsuitable" cases to be counted, only hits and misses.

The remaining static debug counter increments were almost identical, and
they have been consolidated.

The "Fold isArray test" trace message is removed because folding will
already be plainly visible in logs due to the performTransformation().