Check whether block is in nested loop
In considering whether to pull an invariant expression out of a loop,
removeUncertainBlocks needs to verify that blocks that are being
considered will be executed exactly once within the region that contains
them.  If a block is nested within another loop in the region, it might
be executed more or less than one time, and so should be removed from
consideration.

Also removed some stray tab characters.