Define static TR_X86ProcessorInfo in static method
Previously, the static instance of TR_X86ProcessorInfo existed as a
member of OMR::X86::CodeGenerator. However, there was never a
well defined point at which it was zero-initialized. This commit moves
the static instance from being a member to being a variable within a
method of OMR::X86::CodeGenerator. This ensures that the instance is
properly zero-initialized once.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>