AArch64: Implement evaluators for additional masked vector binary operations
This commit implements vmul/vdiv/vmin/vmax evaluators.
`inlineVectorBinaryOp` and `inlineVectorMaskedBinaryOp` are changed to
take an optional function pointer argument and call the pointed function
if it is not null.
Helper functions for vmul/vdiv/vmin/vmax are changed to
take result, first argument and second argument registers as parameters
instead of evaluating nodes inside them.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>