Correct constraint range for iushr and lushr with negative operand
The value propagation handlers for iushr and lushr have special handling
for the cases where the lower bound (L) of the left operand is non-
negative or its upper bound (U) is negative, if the right hand operand's
value (S) has a constant constraint.  If L is non-negative, the handler
sets the range of the result of the operation to be in the range
[L>>S,U>>S]; if U is negative, it sets the range of the result to be
[U>>S,L>>S].  However, the latter range has the bounds reversed, as can
be seen by considering an example.

For instance, if L is 0x80FFFFFF (-2130706433) and U is 0xFFFFFFFF (-1)
and S is 24, [U>>S,L>>S] yields the range [0xFF, 0x80] or [255,128].

Corrected this by computing the result range the same way for both non-
negative ranges and negative ranges.