Fix discrepancy in GC Extensions Base functions
There is a number of functions in the GC Extensions Base defined under
the OMR_GC_MODRON_SCAVENGER flag. However, there are many places where
these functions are called unconditionally without checking this flag.
This flag is always enabled in our compilation environment, which
allows the code to compile. If it is disabled, the code fails to
compile. The solution is to always define these functions and wrap
Scavenger related content inside them.

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>