Improve value propagation handlers for integer truncation
Previously, the result would be unconstrained whenever it was possible
for the child to take on the (sign-extension of) the min or max value of
the result type.

The result would also be unconstrained whenever it was possible for the
child to take on a value outside of the signed range of the result type.
This was particularly unhelpful for a constant child.

Finally, the truncation node would never be constant-folded even when
its result was constrained to be constant.

Now, a range constraint is produced whenever possible. For example, if
the child of i2b is constrained to [511, 513], then the i2b will now be
constrained to [-1, 1]. If OTOH the child is constrained to [383, 384],
then the result will be unconstrained. When the child is constant, it is
always possible to constrain, and the resulting constraint is also
constant. In this case, the node is now folded.

In particular, this allows VP to see the result of an is-array test for
a known class on 64-bit even when the flags are out of signed 32-bit
range, e.g.

    iushr
      iand
        l2i (*)
          lconst 0x8yyyyyyy
        iconst 0x10000
      iconst 16

Previously, failure to constrain l2i would stop propagation there.