Record kind of transformation that is under consideration
Expressions Simplification looks for summation reductions and invariant
expressions that can be pulled out of loops.  Each instance is tracked
by placing the TR::TreeTop of the candidate in a TR::List.  However, in
the process of actually performing the transformation, <some method>
doesn't perform the same checking of a summation reduction as it does
during the initial analysis.  In particular, an invariant expression
that happens to involve an ineg, ixor, isub or iadd might end up being
transformed as if it was a summation reduction.

Fixed this by using a list of candidates that carries a flag indicating
the kind of transformation that can be performed along with the
TR::TreeTop.