Add retry 10 around clone with 1 minute delays
There is a problem on zos where the known_hosts
for github.com cannot be verified. It is intermittent
and can come and go on either machine seemingly randomly.
This changes adds a retry wrapper to the clone which will
retry the clone if it fails up to 10 times. It also adds
a delay of 1 minute in between each try.

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>