AArch64 Mac: Update JBCodeCacheManager to enable jitbuilder
This commit updates JBCodeCacheManager to allocate
the CodeCacheMemorySegment object by malloc instead of using a part of
the mmaped memory segment on AArch64 Mac. The mmaped memory segment is
not both writable and executable at the same time on Apple Silicon.
If the CodeCacheMemorySegment object is allocated on the mmaped segment,
an access violation would occur when updating its members.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>