Clean up and likely speed up printing of node flags
The node flag "print" methods are trivial. They are almost all
one-liners, and their definitions are highly repetitive (not within
definitions, but across them). These methods really only serve as a
noisy and boilerplate-heavy way to specify the flags' names.
Furthermore, each has exactly one use site, so they don't even
meaningfully deduplicate those names in the source code.

Now these methods are inlined into TR_Debug::nodePrintAllFlags(), which
centralizes the choice of names for all flags in trace logs. This most
likely helps the performance of node flag printing on its own, since
previously it was not possible for the build compiler to inline any of
the "print" methods (except for one). Additionally, and in part enabled
by the inlining,
- the repetitive structure is factored out;
- unset flags now skip appending, instead of appending ""; and
- output is generated using appends() instead of appendf("%s", ...).

After inlining, the "print" methods are unused, and they are removed.

This commit does not change the set of flags printed explicitly in
nodePrintAllFlags(), their printed names, or their order. Specifically,
OpenJ9-specific flags are still printed here. Handling of those flags
should be moved to OpenJ9, but that move is not tackled here.