Detect invariant checkcasts and conditionals separately in versioner
Previously, the generic-sounding detectInvariantTrees() handled both
even though it's completely filled with logic specific to conditional
trees. That method is difficult enough to understand and modify without
needing to consider the possibility of checkcasts, and it obscures the
simplicity of what we actually require for checkcasts.

As such, it is renamed detectInvariantConditionals(), and checkcasts are
now handled instead by the new detectInvariantCheckCasts().