Preserve flags in simplifier transformations
In particular, for a constant c > 0, preserve:

- zero/sign flags and cannotOverflow when normalizing iadd to isub, i.e.
  where x+c becomes x-(-c);

- cannotOverflow when rewriting an integer addition (-x)+y to y-x;

- cannotOverflow when normalizing isub to iadd, i.e. where x-c becomes
  x+(-c);

- cannotOverflow when rewriting a negation -(x-y) to y-x;

- cannotOverflow (if possible) when coalescing right-hand-side constants
  for consecutive iadd/isub operations; and

- cannotOverflow (if possible) when extracting ineg from imul, i.e.
  where x*(-c) becomes -(x*c);

In all but the first case, the zero/sign flags are already preserved.

Additionally, in the last case, i.e. when extracting ineg from imul, the
resulting ineg and the new imul are either both zero or both nonzero. If
ineg can't overflow, then neither can the new imul, and any pair of
corresponding nonzero results from imul and ineg have opposite signs.