Make TR::Node::recreate() unset the cannotOverflow flag
This flag is about the operation to be performed, which is changing.
Flags that describe the resulting value can always be kept. There are
likely other flags that should be unset here, but this is a start.

The following are a few examples of cases where the new/modified
operation could overflow even if the original cannot. The first one is
implemented in simplifier. The second will probably not happen, since
the compiler prefers isub, but it is simple and hopefully illustrative.

    imul (cannotOverflow)           ineg
      x                     ==>       imul
      iconst -2                         x
                                        iconst 2

Here, there is (in general) a possibility that x could be 2**30. In that
case, multiplying by -2 gives -2**31 (i.e. MIN) without overflow.
However, if we multiply instead by 2, the true product is 2**31, which
is not representable, so the product overflows and produces MIN,
Furthermore, if the result of imul is MIN, then ineg overflows too.

    isub (cannotOverflow)           iadd
      x                     ==>       x
      iconst MIN                      iconst MIN

For isub not to overflow, we require only that x be negative. The iadd
will have the same result, but because x is negative, it will always
overflow.