Set inlined call site index to -1 for all OSR guards
An OSR guard does not correspond to any inlined call site in particular.
The OSR assumptions are made throughout the method in a generalized way.
Using -1 as the call site index helps to prevent later analyses from
confusing an OSR guard for a virtual guard, since it will not be
possible to find a call that appears to have corresponding BCI.