Avoid analyzing OSR guards as virtual guards in versioner
It used to be that a standalone OSR guard could not target a cold call.
However, virtual guards target cold calls even when they are merged with
OSR guards. Recently, VP learned to remove the virtual part of a virtual
guard despite any merged guards (3d1fec9523d4f9e7). The merged HCR or
OSR guard is promoted to a standalone HCR or OSR guard, and any
standalone OSR guard produced in this way does target a cold call.

It was possible for such an OSR guard to be confused for a virtual
guard, and therefore for loop versioner to try to version it as though
it were a virtual guard, resulting in an assertion failure:

    guardOkForExpr: should not intern OSR guard ...

This happened only very rarely because the OSR guard does not use VFT
test or method test, so in order to confuse it for a virtual guard, the
OSR guard's BCI must correspond to that of the cold call in the same way
as it would for a virtual guard. An OSR guard's inlined call site index
is copied from the BBStart on the taken side, which usually identifies
the caller rather than the callee, but the latter is required in order
to provoke the bug.

Additionally, buildConditionalTree() had a path that excluded HCR guard
but not direct method guard. It shouldn't matter, since neither should
be considered an invariant conditional there, but that path now also
excludes direct method guard for consistency.