Ensure receivers are privatized (if necessary) when versioning guards
Some guard trees, e.g. for nonoverridden guards, do not contain an
occurrence of the receiver. Since it doesn't occur in the guard tree, it
is not automatically privatized by depsForLoopEntryPrep().

Detect this case and explicitly privatize the receiver (if necessary) to
ensure that it cannot vary from one iteration to the next, even in the
presence of a data race. This will also prevent versioning of the guard
whenever privatization is disallowed, i.e. whenever a cold call will
necessarily remain within the hot loop, since the cold call could change
the receiver.